server/requirements.txt
anyio==3.5.0
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.74.1
firebase-admin==5.2.0
gcloud==0.17.0
google-api-core==2.5.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httplib2==0.20.4
httptools==0.3.0
idna==3.3
jws==0.1.3
msgpack==1.0.3
oauth2client==3.0.0
packaging==21.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pydantic==1.9.0
pyparsing==3.0.7
Pyrebase==3.0.27
python-dotenv==0.19.2
python-jwt==2.0.1
PyYAML==6.0
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
